requirements.txt
pillow
vk_api
flask
gunicorn